feat(plx-ecc9bf): refactor ProcedureCloudWatchLogger to general-purpose PlexusCloudWatchLogger
Generalize CloudWatch Logs integration for use across Plexus components
(console chat, procedures, optimizer, etc.).

Changes:
- Created plexus/logging/cloudwatch_logger.py with PlexusCloudWatchLogger
- Renamed parameters: procedure_id → component_name, invocation_run_id → invocation_id
- Added log_category parameter for log group flexibility (default: 'procedures')
- Updated log events: procedure_started/completed → component_started/completed
- Refactored plexus/cli/procedure/cloudwatch_logger.py to use new logger
- Kept procedure-specific factory functions and Tactus patching

Benefits:
- DRY: Single CloudWatch Logs implementation for all Plexus components
- Flexible: Configurable log categories and component names
- Backward compatible: Existing procedure logging continues to work unchanged

Log structure:
- Log group: /plexus/{log_category}/{account_key}
- Run stream: {component_name}/run/{invocation_id}
- LLM context stream: {component_name}/llm-context/{invocation_id}

This enables console chat to use the same CloudWatch logging infrastructure
as procedures, with proper segmentation by component type.
