+++ plexus/logging/__init__.py
@@ -1,0 +1,5 @@
+"""Plexus logging utilities."""
+
+from .cloudwatch_logger import PlexusCloudWatchLogger
+
+__all__ = ["PlexusCloudWatchLogger"]
