+++ project/issues/plx-ecc9bf2d-e353-4740-8ac1-1ff9cd58d1d4.json
@@ -3,7 +3,7 @@
   "title": "Refactor ProcedureCloudWatchLogger to general-purpose PlexusCloudWatchLogger",
   "description": "",
   "type": "task",
-  "status": "open",
+  "status": "closed",
   "priority": 1,
   "assignee": null,
   "creator": null,
@@ -25,7 +25,7 @@
     }
   ],
   "created_at": "2026-05-14T17:51:13.996588Z",
-  "updated_at": "2026-05-14T18:09:31.032438Z",
-  "closed_at": null,
+  "updated_at": "2026-05-14T20:13:19.434222Z",
+  "closed_at": "2026-05-14T20:13:19.434222Z",
   "custom": {}
 }
